Cleaning hpx_fwd.hpp from action forward definitions
